+++ support-frontend/assets/pages/[countryGroupId]/router.tsx
@@ -8,7 +8,7 @@ import {
 import { ClientSideErrorPage } from 'components/ClientSideError';
 import { GuardianHoldingContent } from 'components/serverSideRendered/guardianHoldingContent';
 import { ObserverHoldingContent } from 'components/serverSideRendered/observerHoldingContent';
-import { FeatureSwitchesProvider } from 'contexts/FeatureFlagsContext';
+import { FeatureSwitchesProvider } from 'contexts/FeatureSwitchesContext';
 import { getStudentLandingPageTestConfig } from 'helpers/abTests/studentLandingPageAbTests';
 import { WithCoreWebVitals } from 'helpers/coreWebVitals/withCoreWebVitals';
 import { isSwitchOn } from 'helpers/globalsAndSwitches/globals';
