+++ support-frontend/assets/pages/[countryGroupId]/thankYou.tsx
@@ -1,7 +1,7 @@
 import { storage } from '@guardian/libs';
 import type { SupportRegionId } from '@modules/internationalisation/countryGroup';
 import { BillingPeriod } from '@modules/product/billingPeriod';
-import { useFeatureSwitches } from 'contexts/FeatureFlagsContext';
+import { useFeatureSwitches } from 'contexts/FeatureSwitchesContext';
 import type { AppConfig } from 'helpers/globalsAndSwitches/window';
 import { Country } from 'helpers/internationalisation/classes/country';
 import {
